pyproject.toml
[tool.poetry]
name = "doc-qna"
version = "0.1.0"
description = "having a chat with documents to aquire knowledge quickly"
authors = ["Eric Ballnath"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
python-dotenv = "^1.0.1"
langchain = "^0.2.5"
pypdf2 = "^3.0.1"
chromadb = "^0.5.3"
langchain-community = "^0.2.5"
pypdf = "^4.2.0"
faiss-cpu = "^1.8.0"
google-generativeai = "0.5.4"
langchain-google-genai = "^1.0.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
